requirements.txt
python-kucoin
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
pytimeparse